+++ lib/isolate_cluster.dart
@@ -9,4 +9,5 @@ import 'package:logging/logging.dart';
 part 'messages.dart';
 part 'isolate.dart';
 part 'cluster.dart';
+part 'helper.dart';